Finalizando a customização de exceções e finalizando o projeto.

+++ src/main/java/br/com/desafiocruddeclientes/cruddeclientes/services/ClientService.java
@@ -3,11 +3,15 @@ package br.com.desafiocruddeclientes.cruddeclientes.services;
 import br.com.desafiocruddeclientes.cruddeclientes.dto.ClientDTO;
 import br.com.desafiocruddeclientes.cruddeclientes.entities.Client;
 import br.com.desafiocruddeclientes.cruddeclientes.repositories.ClientRepository;
+import br.com.desafiocruddeclientes.cruddeclientes.services.exceptions.DatabaseException;
+import br.com.desafiocruddeclientes.cruddeclientes.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -18,8 +22,11 @@ public class ClientService {
 
   @Transactional(readOnly = true)
   public ClientDTO findById(Long id) {
-    Client cliente = repository.findById(id).get();
-    return new ClientDTO(cliente);
+    Client client = repository
+      .findById(id)
+      .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado")
+      );
+    return new ClientDTO(client);
   }
 
   @Transactional(readOnly = true)
@@ -39,15 +46,27 @@ public class ClientService {
 
   @Transactional
   public ClientDTO update(Long id, ClientDTO dto) {
-    Client entity = repository.getReferenceById(id);
-    copyDtoToEntity(dto, entity);
-    entity = repository.save(entity);
+    try {
+      Client entity = repository.getReferenceById(id);
+      copyDtoToEntity(dto, entity);
+      entity = repository.save(entity);
 
-    return new ClientDTO(entity);
+      return new ClientDTO(entity);
+    } catch (EntityNotFoundException e) {
+      throw new ResourceNotFoundException("Recurso não encontrado");
+    }
   }
 
+  @Transactional(propagation = Propagation.SUPPORTS)
   public void delete(Long id) {
-    repository.deleteById(id);
+    if (!repository.existsById(id)) {
+      throw new ResourceNotFoundException("Recurso não encontrado");
+    }
+    try {
+      repository.deleteById(id);
+    } catch (DataIntegrityViolationException e) {
+      throw new DatabaseException("Falha de integridade referencial");
+    }
   }
 
   private void copyDtoToEntity(ClientDTO dto, Client entity) {
